CORCI-663 build: Don't ignore scons build error
Issues as that tee_file isn't always set so if scons build
failed, it would get interpreted as a log failure but
PIPESTATUS[1] would be empty to the build would happily pass.
Adding | true as default setting for "tee_file" fixes the
issue.
PIPESTATUS getting reset after first use so save it first

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>