CORCI-734 build: Add publishToRepository step (#58)
Add a publishToRepository step to the public pipeline-lib as a wrapper
for an optional system provided step that handles the actual repository
operations.  This will eventually allow other Jenkins systems to use
our jenkinsfiles and pipeline-lib repository.

Add a daosStackNotifyStatus step to handle what ever notification
is needed for build status.  This expects the parameters for the
GithubNotify Jenkins step.

Signed-off-by: John.E.Malmberg <john.e.malmberg@intel.com>