CORCI-922 build: Use $GIT_COMMIT as the default object to check out
scm.branches contains a branch name, not a commit.  That means that
on Replay, where you want to re-build an older build, scm will build
the branch tip, not the commit being Replayed.

Also in the case quick successive commits to a branch, again since
scm.branches contains the branch name, by the time the first of some
successive commits gets to the point of doing the checkout, the branch
tip might be a newer commit than that being built, in which case the
build ends up bulding a subsequent commit, not the one it should be.

Use $GIT_COMMIT to resolve this issue, as it should always point at
the specific commmit, no the branch tip.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>