Remove later test_tag override
In working on a previous patch, it was missed that subsequent to the
config['test_tag'] processing and combining with the cluster size
restricting tags, into result['test_tag'], a subsequent override of the
result['test_tag'] with the bare config['test_tag'] was being done.

Remove that.

Skip-PR-comments: true

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>