CORCI-1081 ci: Changes to increase CI throughput
Change PRs to only run CentOS7 on VMs.

Landings and timed builds are moved down to the same priority as PRs.

Run build and test stages on an empty commit message with PR-repos* set.
Because there are no code changes, these commits will pass the doc-only
test and not run build and test, but the PR-repos* pragmas tell us that
this is a commit being used to test updates to a dependency build.

Skip UT stage on Quick-build (which is implied with Quick-Functional).

Fix UT not skipping with RPM-test-version:.

Fix Pre-build and skipping for quick-functional.

Skip-PR-comments: true

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>