Revert minor RHEL build changes (#246)
As they aren't really needed and just unnecessarily complicates things.

Also don't do any pre-build if testing prev. built RPMs since that
source is already in the can.

Add a submodule status to display the commit of the packaging repo being
used.  Typically to verify that we always have packaging's master.

Add a step to the RPM build stages to run rpmlint on the result.  This
should ideally be moved to separate stages but we either have to stash
all of the artifacts (that we are archiving already) or figure out how
to fetch the archived artifacts.  This is all just advisory only for now.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>