Revert "DAOS-623 build: Update docker cachebust to be weekly. (#292)"
This reverts commit 62bc371d9c568cab1aaa4dadb3e755ad993c1bce.

CACHEBUST was always intended to be an absolute cache buster.  It was
not meant to represent tasks that were subsequently decided should be
done weekly instead of done always.  CB0 is in fact exactly that -- a
weekly cache buster and should have been used where CACHEBUST was
repurposed to be weekly.

There are, and always will be tasks which we want to happen on every
Docker build.  These are forced with CACHEBUST.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>