DAOS-13527 test: Support default tag assignment through Jenkinsfile
The test tags to use the functional test stages should be driven by the
method through which the user has started the build. If the build has
been built with parameters or started by a timer then the parameters of
that build should drive the test tag selection. If the user pushes a PR
then any tags defined via commit pragmas should be used followed by
the any build parameter tags.

Signed-off-by: Phil Henderson <phillip.henderson@intel.com>