DAOS-623 test: Fix link to Jenkins test results.
Test results have a optional class name  which defaults
to (root) however the Github comments on test failures link
to this default class.  This means that for tests which define
a class but fail the link can either be invalid or point to
other, un-related tests which might not have failed.

Signed-off-by: Ashley Pittman <ashley.m.pittman@intel.com>