Get hardware distro version from branch/distro map
Don't need the downstream branch matching in distroVersion() since it
will always be just 1 or 2 commits ahead of the base branch we are
looking for and can be found by walking back to find the base branch
eliminating the need to maintain an RE for it.

If not determined by the pipeline job (in env.BASE_BRANCH_NAME) find the
base branch by walking backward until we find a valid base branch.  Use
this when trying to use the branch/distro/version map.

Actually add the pragmas to the environment in pragmasToEnv() in
addition to returning the string that was added.

Sync the skipping logic for the RPMs test stages.

Add a Test-downstream-test pragma to allow skipping testing in the
downstream jobs.

Make sure to fetch any updates in the downstream testing job setup.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>