SRE-3104 ci: Use DAOS specific HTTPS_PROXY variable (#470)
SRE-3104 ci: Use DAOS_HTTPS_PROXY environment variable instead of HTTPS_PROXY

The HTTPS_PROXY variable set in Jenkins may force any shell script to use a proxy server.
This is not expected, as we want to avoid using a proxy unless it is explicitly needed by
the DAOS tests or the build process.
In details:
* Pass HTTPS_PROXY explicitly to the build script.
* Pass HTTPS_PROXY to post_provision_config.sh
* Pass HTTPS_PROXY to Docekr build command
* pip3 install requires proxy to access internet
* pip3 install requires explicitly pass proxy parameter with `--proxy`

Signed-off-by: Tomasz Gromadzki <tomasz.gromadzki@hpe.com>